#4 / 패키지 이름 변경 sopt.uni_aos => sopt.uni
