+++ app/build.gradle
@@ -12,11 +12,11 @@ Properties properties = new Properties()
 properties.load(project.rootProject.file('local.properties').newDataInputStream())
 
 android {
-    namespace 'sopt.uni_aos'
+    namespace 'sopt.uni'
     compileSdk 33
 
     defaultConfig {
-        applicationId "sopt.uni_aos"
+        applicationId "sopt.uni"
         minSdk 28
         targetSdk 33
         versionCode 1
