+++ app/src/main/java/sopt/uni/data/entity/ExampleEntity.kt
@@ -1,0 +1,5 @@
+package sopt.uni.data.entity
+
+data class ExampleEntity(
+    val id: Int,
+)
