+++ app/src/main/java/sopt/uni/data/repository/example/ExampleRepositoryImpl.kt
@@ -1,0 +1,9 @@
+package sopt.uni.data.repository.example
+
+import sopt.uni.data.entity.ExampleEntity
+import sopt.uni.data.source.remote.request.ExampleRequestDto
+
+class ExampleRepositoryImpl() : ExampleRepository {
+    override suspend fun postExample(exampleRequestDto: ExampleRequestDto): Result<ExampleEntity> =
+        Result.success(ExampleEntity(1))
+}
