+++ app/src/main/java/sopt/uni/data/source/remote/request/ExampleRequestDto.kt
@@ -1,0 +1,12 @@
+package sopt.uni.data.source.remote.request
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ExampleRequestDto(
+    @SerialName("id")
+    val id: String,
+    @SerialName("type")
+    val type: String,
+)
