+++ app/src/main/java/sopt/uni/data/source/remote/response/ExampleResponseDto.kt
@@ -1,0 +1,10 @@
+package sopt.uni.data.source.remote.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ExampleResponseDto(
+    @SerialName("id")
+    val id: Int,
+)
