+++ app/src/main/java/sopt/uni/di/RepositoryModule.kt
@@ -1,0 +1,9 @@
+package sopt.uni.di
+
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule
