+++ app/src/main/java/sopt/uni/di/ServiceModule.kt
@@ -1,0 +1,9 @@
+package sopt.uni.di
+
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object ServiceModule
