+++ app/src/main/java/sopt/uni/presentation/MainActivity.kt
@@ -1,0 +1,15 @@
+package sopt.uni.presentation
+
+import android.os.Bundle
+import dagger.hilt.android.AndroidEntryPoint
+import sopt.uni.R
+import sopt.uni.databinding.ActivityMainBinding
+import sopt.uni.util.binding.BindingActivity
+
+@AndroidEntryPoint
+class MainActivity : BindingActivity<ActivityMainBinding>(R.layout.activity_main) {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+    }
+}
