+++ app/src/main/java/sopt/uni/util/EventObserver.kt
@@ -1,0 +1,11 @@
+package sopt.uni.util
+
+import androidx.lifecycle.Observer
+
+class EventObserver<T>(private val onEventUnhandledContent: (T) -> Unit) : Observer<Event<T>> {
+    override fun onChanged(event: Event<T>) {
+        event?.getContentIfNotHandled()?.let { value ->
+            onEventUnhandledContent(value)
+        }
+    }
+}
