+++ app/src/main/java/sopt/uni/util/binding/BindingActivity.kt
@@ -1,0 +1,19 @@
+package sopt.uni.util.binding
+
+import android.os.Bundle
+import androidx.annotation.LayoutRes
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+
+abstract class BindingActivity<T : ViewDataBinding>(
+    @LayoutRes private val layoutRes: Int
+) : AppCompatActivity() {
+    protected lateinit var binding: T
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = DataBindingUtil.setContentView(this, layoutRes)
+        binding.lifecycleOwner = this
+    }
+}
