+++ app/src/main/java/sopt/uni/util/extension/ActivityExt.kt
@@ -1,0 +1,9 @@
+package sopt.uni.util.extension
+
+import android.app.Activity
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+
+inline fun <reified T : Activity> AppCompatActivity.startActivity(block: Intent.() -> Unit = {}) {
+    startActivity(Intent(this, T::class.java).apply(block))
+}
