+++ app/src/main/java/sopt/uni/util/extension/ContextExt.kt
@@ -1,0 +1,8 @@
+package sopt.uni.util.extension
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.showToast(msg: String) {
+    Toast.makeText(this, msg, Toast.LENGTH_SHORT).show()
+}
