--- app/src/main/java/sopt/uni_aos/UniApplication.kt
@@ -1,16 +1,0 @@
-package sopt.uni_aos
-
-import android.app.Application
-import androidx.appcompat.app.AppCompatDelegate
-import dagger.hilt.android.HiltAndroidApp
-import timber.log.Timber
-
-@HiltAndroidApp
-class UniApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-
-        if (BuildConfig.DEBUG) Timber.plant(Timber.DebugTree())
-        AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-    }
-}
