--- app/src/main/java/sopt/uni_aos/data/entity/ExampleEntity.kt
@@ -1,5 +1,0 @@
-package sopt.uni_aos.data.entity
-
-data class ExampleEntity(
-    val id: Int,
-)
