--- app/src/main/java/sopt/uni_aos/data/repository/example/ExampleRepositoryImpl.kt
@@ -1,9 +1,0 @@
-package sopt.uni_aos.data.repository.example
-
-import sopt.uni_aos.data.entity.ExampleEntity
-import sopt.uni_aos.data.source.remote.request.ExampleRequestDto
-
-class ExampleRepositoryImpl() : ExampleRepository {
-    override suspend fun postExample(exampleRequestDto: ExampleRequestDto): Result<ExampleEntity> =
-        Result.success(ExampleEntity(1))
-}
