--- app/src/main/java/sopt/uni_aos/data/source/local/Example.kt
@@ -1,3 +1,0 @@
-package sopt.uni_aos.data.source.local
-
-// 로컬 데이터 소스 어쩌구 저쩌구
