--- app/src/main/java/sopt/uni_aos/data/source/remote/request/ExampleRequestDto.kt
@@ -1,12 +1,0 @@
-package sopt.uni_aos.data.source.remote.request
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ExampleRequestDto(
-    @SerialName("id")
-    val id: String,
-    @SerialName("type")
-    val type: String,
-)
