--- app/src/main/java/sopt/uni_aos/data/source/remote/response/ExampleResponseDto.kt
@@ -1,10 +1,0 @@
-package sopt.uni_aos.data.source.remote.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ExampleResponseDto(
-    @SerialName("id")
-    val id: Int,
-)
