--- app/src/main/java/sopt/uni_aos/di/RepositoryModule.kt
@@ -1,9 +1,0 @@
-package sopt.uni_aos.di
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class RepositoryModule
