--- app/src/main/java/sopt/uni_aos/di/ServiceModule.kt
@@ -1,9 +1,0 @@
-package sopt.uni_aos.di
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-object ServiceModule
