--- app/src/main/java/sopt/uni_aos/util/Event.kt
@@ -1,17 +1,0 @@
-package sopt.uni_aos.util
-
-open class Event<out T>(private val content: T) {
-    var hasBeenHandled = false
-        private set
-
-    fun getContentIfNotHandled(): T? {
-        return if (hasBeenHandled) {
-            null
-        } else {
-            hasBeenHandled = true
-            content
-        }
-    }
-
-    fun peekContent(): T = content
-}
