--- app/src/main/java/sopt/uni_aos/util/EventObserver.kt
@@ -1,11 +1,0 @@
-package sopt.uni_aos.util
-
-import androidx.lifecycle.Observer
-
-class EventObserver<T>(private val onEventUnhandledContent: (T) -> Unit) : Observer<Event<T>> {
-    override fun onChanged(event: Event<T>) {
-        event?.getContentIfNotHandled()?.let { value ->
-            onEventUnhandledContent(value)
-        }
-    }
-}
