--- app/src/main/java/sopt/uni_aos/util/binding/BindingActivity.kt
@@ -1,19 +1,0 @@
-package sopt.uni_aos.util.binding
-
-import android.os.Bundle
-import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-
-abstract class BindingActivity<T : ViewDataBinding>(
-    @LayoutRes private val layoutRes: Int
-) : AppCompatActivity() {
-    protected lateinit var binding: T
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, layoutRes)
-        binding.lifecycleOwner = this
-    }
-}
