--- app/src/main/java/sopt/uni_aos/util/binding/BindingAdapter.kt
@@ -1,13 +1,0 @@
-package sopt.uni_aos.util.binding
-
-import android.widget.ImageView
-import androidx.databinding.BindingAdapter
-import coil.load
-
-object BindingAdapter {
-    @JvmStatic
-    @BindingAdapter("setImage")
-    fun setImage(view: ImageView, imgUrl: String?) {
-        view.load(imgUrl)
-    }
-}
