--- app/src/main/java/sopt/uni_aos/util/binding/BindingFragment.kt
@@ -1,33 +1,0 @@
-package sopt.uni_aos.util.binding
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-import androidx.fragment.app.Fragment
-import sopt.uni_aos.R
-
-abstract class BindingFragment<T : ViewDataBinding>(
-    @LayoutRes private val layoutRes: Int
-) : Fragment() {
-    private var _binding: T? = null
-    protected val binding get() = _binding ?: error(getString(R.string.binding_error))
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = DataBindingUtil.inflate(inflater, layoutRes, container, false)
-        binding.lifecycleOwner = viewLifecycleOwner
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
