--- app/src/main/java/sopt/uni_aos/util/extension/ActivityExt.kt
@@ -1,9 +1,0 @@
-package sopt.uni_aos.util.extension
-
-import android.app.Activity
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-
-inline fun <reified T : Activity> AppCompatActivity.startActivity(block: Intent.() -> Unit = {}) {
-    startActivity(Intent(this, T::class.java).apply(block))
-}
