--- app/src/main/java/sopt/uni_aos/util/extension/ContextExt.kt
@@ -1,8 +1,0 @@
-package sopt.uni_aos.util.extension
-
-import android.content.Context
-import android.widget.Toast
-
-fun Context.showToast(msg: String) {
-    Toast.makeText(this, msg, Toast.LENGTH_SHORT).show()
-}
