--- app/src/main/java/sopt/uni_aos/util/extension/ItemDiffCallback.kt
@@ -1,12 +1,0 @@
-package sopt.uni_aos.util.extension
-
-import androidx.recyclerview.widget.DiffUtil
-
-class ItemDiffCallback<T : Any>(
-    val onItemsTheSame: (T, T) -> Boolean,
-    val onContentsTheSame: (T, T) -> Boolean
-) : DiffUtil.ItemCallback<T>() {
-    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean = onItemsTheSame(oldItem, newItem)
-    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean =
-        onContentsTheSame(oldItem, newItem)
-}
