--- app/src/main/java/sopt/uni_aos/util/extension/ParcelableExt.kt
@@ -1,16 +1,0 @@
-package sopt.uni_aos.util.extension
-
-import android.content.Intent
-import android.os.Build
-import android.os.Bundle
-import android.os.Parcelable
-
-inline fun <reified T : Parcelable> Intent.parcelable(key: String): T? = when {
-    Build.VERSION.SDK_INT >= 33 -> getParcelableExtra(key, T::class.java)
-    else -> @Suppress("DEPRECATION") getParcelableExtra(key) as? T
-}
-
-inline fun <reified T : Parcelable> Bundle.parcelable(key: String): T? = when {
-    Build.VERSION.SDK_INT >= 33 -> getParcelable(key, T::class.java)
-    else -> @Suppress("DEPRECATION") getParcelable(key) as? T
-}
